+++ packages/localnet/src/deployOpts.ts
@@ -1,0 +1,4 @@
+export const deployOpts = {
+  gasPrice: 10000000000,
+  gasLimit: 6721975,
+};
